requirements.txt
# Gmail API and OAuth2 authentication
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# LLM integration
openai==1.97.1

# CLI and utilities
colorama==0.4.6
python-dateutil==2.8.2

# Email parsing and analysis
email-validator==2.1.0
beautifulsoup4==4.12.2
html2text==2020.1.16

# Data processing
pandas==2.1.4
numpy==1.24.4

# Security and environment
cryptography==41.0.8
